configs: now each modules except cfg.py don't have side effect anymore, the configs loading are done in configs.cfg module
